adds a timestamp to each entry in scriptout
The goal is to enable a reliable source for analyzing user input, which is not possible with the current on_key method, since it doesn't reflect the actual keys pressed.
It would be better if I added another option instead of overwriting -w, but for personal usage is fine for now.